Add delete url

+++ src/main/java/com/theruzil/shorter/repository/UrlRepository.java
@@ -5,9 +5,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface UrlRepository extends CrudRepository<Url, Long> {
+public interface UrlRepository extends CrudRepository<Url, Integer> {
     Url findById(long id);
     List<Url> findByFullUrl(String fullUrl);
-
     List<Url> findAll();
 }
